feat: ask before delete, with command line option to override

+++ docker_registry_purger/purger.py
@@ -4,6 +4,7 @@ from natsort import natsorted
 import re
 
 from .semver import Semver
+from .utils import query_yes_no
 
 
 logger = daiquiri.getLogger(__name__)
@@ -22,9 +23,10 @@ class Item:
 
 
 class Purger:
-    def __init__(self, registry, strategies=[], dry_run=True):
+    def __init__(self, registry, strategies=[], proceed=None, dry_run=False):
         self.registry = registry
         self.strategies = strategies
+        self.proceed = proceed
         self.dry_run = dry_run
 
     def add_strategy(self, strategy):
@@ -32,11 +34,11 @@ class Purger:
 
     def run(self, repos):
         # Phase/pass 1: gather items
-        logger.info('Phase 1: gather items...')
+        logger.info('Phase 1: gather items ...')
         repo_items = {repo: self.gather_items(repo) for repo in repos}
 
         # Phase 2: mark items for deletion
-        logger.info('Phase 2: mark items for deletion...')
+        logger.info('Phase 2: mark items for deletion ...')
         for items in repo_items.values():
             self.mark_items(items)
 
@@ -49,9 +51,12 @@ class Purger:
         logger.info('Keeping %d undecided items', sum(1 for item in items if item.keep is None))
         logger.debug('Undecided items: %s', ', '.join((f'{item.repo}:{item.tag or item.digest}' for item in items if item.keep is None)))
 
-        # Phase 3: delete marked items
-        logger.info('Phase 3: delete marked items...')
-        self.delete_marked_items(items)
+        if self.proceed == False:
+            logger.info('Skipping deletion because proceed=False')
+        elif self.proceed == True or query_yes_no('\nProceed with delete?'):
+            # Phase 3: delete marked items
+            logger.info('Phase 3: delete marked items ...')
+            self.delete_marked_items(items)
 
     def gather_items(self, repo):
         logger.info('Checking repository <%s>', repo)
@@ -87,5 +92,6 @@ class Purger:
 
     def delete_marked_item(self, item):
         if item.keep == False:
+            logger.info('%sDeleting %s:%s ...', '[dry run]' if self.dry_run else '', item.repo, item.digest or item.tag)
             if not self.dry_run:
                 self.registry.delete_digest_or_tag(item.repo, item.digest, item.tag)
